requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
statsmodels==0.14.0
joblib==1.3.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# API and HTTP
requests==2.31.0
httpx==0.25.2

# Kafka
kafka-python==2.0.2

# Data processing
python-multipart==0.0.6
pydantic==2.5.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Monitoring
prometheus-client==0.19.0

# ElasticSearch
elasticsearch==8.11.0
